Refactor build_repo_config
